feat: refactor advanced permissions handling to use dedicated feature flags module
