+++ shared/lib/environment.test.ts
@@ -1,7 +1,5 @@
 import { ENVIRONMENT } from '../../development/build/constants';
 import {
-  ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG,
-  getEnabledAdvancedPermissions,
   getIsPerpsIncludedInBuild,
   getIsPasskeyFeatureEnabled,
   getIsAssetsUnifiedStateIncludedInBuild,
@@ -34,144 +32,6 @@ describe('isProduction', () => {
   it('should return false when ENVIRONMENT is "testing"', () => {
     process.env.METAMASK_ENVIRONMENT = ENVIRONMENT.TESTING;
     expect(isProduction()).toBe(false);
-  });
-});
-
-describe('getEnabledAdvancedPermissions', () => {
-  let originalGatorEnabledPermissionTypes: string | undefined;
-
-  beforeAll(() => {
-    originalGatorEnabledPermissionTypes =
-      process.env.GATOR_ENABLED_PERMISSION_TYPES;
-  });
-
-  afterAll(() => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      originalGatorEnabledPermissionTypes;
-  });
-
-  it('should return an empty array when GATOR_ENABLED_PERMISSION_TYPES is not set', () => {
-    delete process.env.GATOR_ENABLED_PERMISSION_TYPES;
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([]);
-  });
-
-  it('should return an empty array when GATOR_ENABLED_PERMISSION_TYPES is an empty string', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES = '';
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([]);
-  });
-
-  it('should parse comma-separated values correctly', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,native-token-periodic,erc20-token-stream';
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([
-      'native-token-stream',
-      'native-token-periodic',
-      'erc20-token-stream',
-    ]);
-  });
-
-  it('should filter out empty strings from the result', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,,erc20-token-stream';
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([
-      'native-token-stream',
-      'erc20-token-stream',
-    ]);
-  });
-
-  it('should handle a single permission type', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES = 'native-token-stream';
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([
-      'native-token-stream',
-    ]);
-  });
-
-  it('should use the remote flag when it is set', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES = 'native-token-stream';
-
-    expect(
-      getEnabledAdvancedPermissions({
-        remoteFeatureFlags: {
-          [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]: {
-            permissions: [
-              'native-token-stream',
-              'native-token-periodic',
-              'erc20-token-periodic',
-              'erc20-token-stream',
-              'erc20-token-revocation',
-              'native-token-allowance',
-              'erc20-token-allowance',
-            ],
-          },
-        },
-      }),
-    ).toStrictEqual([
-      'native-token-stream',
-      'native-token-periodic',
-      'erc20-token-periodic',
-      'erc20-token-stream',
-      'erc20-token-revocation',
-    ]);
-  });
-
-  it('should ignore permission types that are not implemented in the extension', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,native-token-allowance';
-
-    expect(getEnabledAdvancedPermissions()).toStrictEqual([
-      'native-token-stream',
-    ]);
-
-    expect(
-      getEnabledAdvancedPermissions({
-        remoteFeatureFlags: {
-          [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]: {
-            permissions: [
-              'native-token-stream',
-              'native-token-allowance',
-              'erc20-token-allowance',
-            ],
-          },
-        },
-      }),
-    ).toStrictEqual(['native-token-stream']);
-  });
-
-  it('should allow the remote flag to disable all permission types', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,erc20-token-stream';
-
-    expect(
-      getEnabledAdvancedPermissions({
-        remoteFeatureFlags: {
-          [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]: {
-            permissions: [],
-          },
-        },
-      }),
-    ).toStrictEqual([]);
-  });
-
-  it('should fall back to the environment when the remote flag is missing', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,erc20-token-stream';
-
-    expect(
-      getEnabledAdvancedPermissions({ remoteFeatureFlags: {} }),
-    ).toStrictEqual(['native-token-stream', 'erc20-token-stream']);
-  });
-
-  it('should fall back to the environment when the remote flag shape is invalid', () => {
-    process.env.GATOR_ENABLED_PERMISSION_TYPES =
-      'native-token-stream,erc20-token-stream';
-
-    expect(
-      getEnabledAdvancedPermissions({
-        remoteFeatureFlags: {
-          [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]: true,
-        },
-      }),
-    ).toStrictEqual(['native-token-stream', 'erc20-token-stream']);
   });
 });
