+++ shared/lib/gator-permissions/index.ts
@@ -2,3 +2,4 @@ export * from './gator-permissions-utils';
 export * from './time-utils';
 export * from './numbers-utils';
 export * from './address-rule-utils';
+export * from './feature-flags';
