+++ test/e2e/feature-flags/feature-flag-registry.test.ts
@@ -1,7 +1,7 @@
 import {
   ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG,
   getEnabledAdvancedPermissions,
-} from '../../../shared/lib/environment';
+} from '../../../shared/lib/gator-permissions/feature-flags';
 import {
   FEATURE_FLAG_REGISTRY,
   FeatureFlagStatus,
@@ -51,46 +51,60 @@ describe('Feature Flag Registry', () => {
     });
 
     it('registers the enabled advanced permissions remote flag', () => {
+      const originalGatorEnabledPermissionTypes =
+        process.env.GATOR_ENABLED_PERMISSION_TYPES;
+      process.env.GATOR_ENABLED_PERMISSION_TYPES =
+        'native-token-stream,native-token-periodic,erc20-token-stream,erc20-token-periodic,erc20-token-revocation';
+
       const entry =
         FEATURE_FLAG_REGISTRY[ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG];
-      if (!entry) {
-        throw new Error(
-          `${ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG} is not registered`,
-        );
-      }
+      try {
+        if (!entry) {
+          throw new Error(
+            `${ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG} is not registered`,
+          );
+        }
 
-      expect(entry).toStrictEqual({
-        name: ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG,
-        type: FeatureFlagType.Remote,
-        inProd: true,
-        productionDefault: {
-          permissions: [
-            'native-token-stream',
-            'native-token-periodic',
-            'erc20-token-stream',
-            'erc20-token-periodic',
-            'erc20-token-revocation',
-            'native-token-allowance',
-            'erc20-token-allowance',
-          ],
-        },
-        status: FeatureFlagStatus.Active,
-      });
-
-      expect(
-        getEnabledAdvancedPermissions({
-          remoteFeatureFlags: {
-            [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]:
-              entry.productionDefault,
+        expect(entry).toStrictEqual({
+          name: ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG,
+          type: FeatureFlagType.Remote,
+          inProd: true,
+          productionDefault: {
+            permissions: [
+              'native-token-stream',
+              'native-token-periodic',
+              'erc20-token-stream',
+              'erc20-token-periodic',
+              'erc20-token-revocation',
+              'native-token-allowance',
+              'erc20-token-allowance',
+            ],
           },
-        }),
-      ).toStrictEqual([
-        'native-token-stream',
-        'native-token-periodic',
-        'erc20-token-stream',
-        'erc20-token-periodic',
-        'erc20-token-revocation',
-      ]);
+          status: FeatureFlagStatus.Active,
+        });
+
+        expect(
+          getEnabledAdvancedPermissions({
+            remoteFeatureFlags: {
+              [ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG]:
+                entry.productionDefault,
+            },
+          }),
+        ).toStrictEqual([
+          'native-token-stream',
+          'native-token-periodic',
+          'erc20-token-stream',
+          'erc20-token-periodic',
+          'erc20-token-revocation',
+        ]);
+      } finally {
+        if (originalGatorEnabledPermissionTypes === undefined) {
+          delete process.env.GATOR_ENABLED_PERMISSION_TYPES;
+        } else {
+          process.env.GATOR_ENABLED_PERMISSION_TYPES =
+            originalGatorEnabledPermissionTypes;
+        }
+      }
     });
   });
