+++ test/e2e/feature-flags/feature-flag-registry.ts
@@ -17,7 +17,7 @@
  */
 
 import type { Json } from '@metamask/utils';
-import { ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG } from '../../../shared/lib/environment';
+import { ENABLED_ADVANCED_PERMISSIONS_FEATURE_FLAG } from '../../../shared/lib/gator-permissions/feature-flags';
 import { getBooleanFeatureFlag } from '../../../shared/lib/remote-feature-flag-utils';
 
 // ============================================================================
